Simple reduction to get WHNF.
Q: Can thunks be evaluated without the stack running out?
A: Well, once you have a linked data structure, you don't need to run
out of the stack. Of course, the state still requires a recursive data
structure to act like the stack, but the chain of thunks is already the
necessary recursive data structure.